LPS-86069 Move interfaces exposing Nimbus into internal
