+++ modules/apps/portal-security-sso/portal-security-sso-openid-connect-impl/src/main/java/com/liferay/portal/security/sso/openid/connect/internal/OpenIdConnectMetadataFactory.java
@@ -12,18 +12,24 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.openid.connect;
+package com.liferay.portal.security.sso.openid.connect.internal;
 
 import aQute.bnd.annotation.ProviderType;
 
-import java.util.Collection;
+import com.liferay.portal.security.sso.openid.connect.OpenIdConnectServiceException;
+
+import com.nimbusds.openid.connect.sdk.op.OIDCProviderMetadata;
+import com.nimbusds.openid.connect.sdk.rp.OIDCClientMetadata;
 
 /**
- * @author Thuong Dinh
+ * @author Edward C. Han
  */
 @ProviderType
-public interface OpenIdConnectProviderRegistry {
+public interface OpenIdConnectMetadataFactory {
 
-	public Collection<String> getOpenIdConnectProviderNames();
+	public OIDCClientMetadata getOIDCClientMetadata();
+
+	public OIDCProviderMetadata getOIDCProviderMetadata()
+		throws OpenIdConnectServiceException.ProviderException;
 
 }
